Allow Github token authorization for all API versions
The current master branch allows authorization via the GITHUB_TOKEN
option only for v1 and v2. The dev branch contains a fix for this,

https://github.com/eduardoboucas/staticman/pull/405/commits/1f31ac827f13d9aaf94030bf4417b361e79242b7

but the pull request has been open for quite some time,

https://github.com/eduardoboucas/staticman/pull/405

Unfortunately, the dev branch does not seem to work out of the box
for me, so I will just use this small patch to make Staticman work
for me for the time being.